requirements.txt
flask
gunicorn
python-dotenv
PyPDF2
reportlab

langchain
langchain-core
langchain-community
langchain-groq